03-ai-podcast-generator/requirements.txt

transformers>=4.45.0,<5.0.0
torch>=2.0.0
accelerate>=0.24.0


streamlit>=1.28.0
langchain-groq>=0.1.0
langgraph>=0.1.0


cloudscraper>=1.2.0
beautifulsoup4>=4.12.0
requests>=2.31.0


gtts>=2.3.0


python-dotenv>=1.0.0
pandas>=2.0.0
typing-extensions>=4.8.0